chore: refactor archipelago into a testable component
