+++ src/components.ts
@@ -4,15 +4,10 @@ import { IConfigComponent, ILoggerComponent } from '@well-known-components/inter
 import { createLogComponent } from '@well-known-components/logger'
 import { createFetchComponent } from './ports/fetch'
 import { createMetricsComponent } from '@well-known-components/metrics'
-import { AppComponents, GlobalContext } from './types'
+import { AppComponents, GlobalContext, ArchipelagoComponent } from './types'
 import { metricDeclarations } from './metrics'
-import { ArchipelagoComponent } from './controllers/ArchipelagoController'
+import { ArchipelagoController } from './controllers/ArchipelagoController'
 import { createNatsComponent } from '@well-known-components/nats-component'
-
-export type ArchipelagoComponents = {
-  config: IConfigComponent
-  logs: ILoggerComponent
-}
 
 async function getLivekitConf(config: IConfigComponent) {
   const url = await config.requireString('LIVEKIT_URL')
@@ -37,16 +32,17 @@ async function getWsRoomServiceConf(config: IConfigComponent) {
   return { url, secret }
 }
 
-export async function createArchipelagoComponent(components: ArchipelagoComponents): Promise<ArchipelagoComponent> {
-  const { config } = components
-
+export async function createArchipelagoComponent(
+  config: IConfigComponent,
+  logs: ILoggerComponent
+): Promise<ArchipelagoComponent> {
   const flushFrequency = await config.requireNumber('ARCHIPELAGO_FLUSH_FREQUENCY')
   const joinDistance = await config.requireNumber('ARCHIPELAGO_JOIN_DISTANCE')
   const leaveDistance = await config.requireNumber('ARCHIPELAGO_LEAVE_DISTANCE')
   const maxPeersPerIsland = await config.requireNumber('ARCHIPELAGO_MAX_PEERS_PER_ISLAND')
   const workerSrcPath = await config.getString('ARCHIPELAGO_WORKER_SRC_PATH')
 
-  const controller = new ArchipelagoComponent({
+  const controller = new ArchipelagoController({
     flushFrequency,
     archipelagoParameters: {
       joinDistance,
@@ -56,7 +52,7 @@ export async function createArchipelagoComponent(components: ArchipelagoComponen
       wsRoomService: await getWsRoomServiceConf(config)
     },
     workerSrcPath,
-    components
+    components: { logs }
   })
 
   return controller
@@ -72,7 +68,7 @@ export async function initComponents(): Promise<AppComponents> {
   const fetch = await createFetchComponent()
   const metrics = await createMetricsComponent(metricDeclarations, { server, config })
   const nats = await createNatsComponent({ config, logs })
-  const archipelago = await createArchipelagoComponent({ config, logs })
+  const archipelago = await createArchipelagoComponent(config, logs)
 
   return {
     config,
